### 2016-12-29 #149 @whedon
[PRE REVIEW]: hmds: An R Package for Heuristic High and Multi Dimensional Scaling
**Submitting author:** @jirotubuyaki (<a href="http://orcid.org/0000-0001-6871-7727">Masashi Okada</a>)
**Repository:** <a href="https://github.com/jirotubuyaki/hmds" target ="_blank">https://github.com/jirotubuyaki/hmds</a>
**Version:** version 1.0
