### 2016-12-29 #149 @arfon
@noamross - is this a good candidate for rOpenSci?
