### 2016-12-30 #149 @arfon
> I watch the full feed of package submissions, and will chime in if I see candidates for dual-submission.

👍  OK thanks @noamross. Apologies for the noise 😄
