### 2017-01-25 #149 @jirotubuyaki
@arfon
Thank you for sending a message to me. This package is made by Rcpp(R & C++).
I do not have a good candidate. Would you chose a reviewer for me.
