### 2017-03-19 #214 @whedon
[REVIEW]: hmds: An R Package for Heuristic High and Multi Dimensional Scaling
**Submitting author:** @jirotubuyaki (<a href="http://orcid.org/0000-0001-6871-7727">Masashi Okada</a>)
**Repository:** <a href="https://github.com/jirotubuyaki/hmds" target ="_blank">https://github.com/jirotubuyaki/hmds</a>
**Version:** version 1.0

- [x] **Version:** Does the release version given match the GitHub release (version 1.0)?
