### 2017-04-22 #214 @kevinushey
I've finished my initial review for now -- there's a number of things that still need to be finished before a full review can be conducted; that is, I don't think the paper is ready yet. Some comments:

>  Functionality: Have the functional claims of the software been confirmed?

Currently, the package / paper does not outline other competing methods for multidimensional scaling (in R or otherwise). A discussion of how this package compares to other approaches to dimensional reduction would be warranted. For example, what solutions already exist for multidimensional scaling in R, and what particular niche does this package fill? I'm not as familiar with the landscape here but a discussion of these in the paper + a comparison in the package vignettes etc. would be helpful.

>  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Somewhat related to the above -- what are the competing solutions already available, and what niche does this package fill exactly?

>  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Currently, I only see one example usage in the package -- this could be expanded.

>  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

There are currently no automated tests -- some test should be added to ensure that the software behaves well with unexpected inputs.

>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Currently, contributions are currently requested through e-mail; it would likely be more productive to accept contributions through (for example) GitHub issues, and supply a `CONTRIBUTING.md` to indicate what's expected from contributions.
