### 2017-06-15 #214 @jirotubuyaki
I am sorry for being late to improve my software.
I plan some improvement for my GitHub repository.
