+++ inst/doc/my-vignette.Rmd
@@ -71,5 +71,4 @@ The heuristic for Multi Dimensional Scaling is described and explain how to use.
 ## Acknowledgments
 This activity would not have been possible without the support of my family and friends. To my family, thank you for lots of encouragement for me and inspiring me to follow my dreams. I am especially grateful to my parents, who supported me all aspects.  
 
-
 ## References
