### 2018-06-15 #214 @arfon
Hi @jirotubuyaki - how are you getting along here? Are you still interested in pursuing this publication in JOSS?
