### 2018-07-16 #214 @jirotubuyaki
I would like to give up for submit my paper.
I must learn many things for writing papers.
Thank you very much.
